+++ game/bishop.py
@@ -1,20 +1,18 @@
 from game.piece import Piece
 
-
-class Rook(Piece):
+class Bishop(Piece):
     def __init__(self, color):
         super().__init__(color)
-        self.__type__ = "ROOK" 
+        self.__type__ = "BISHOP"
 
     def permited_move(self, from_row, from_col, to_row, to_col, board):
-        if to_row == from_row and to_col != from_col:
+        if abs(to_row - from_row) == abs(to_col - from_col): # :/  :?
             return True
-        elif to_col == from_col and to_row != from_row:
-            return True
-        return False
-    
+        else:
+            return False
+        
     def show(self):
         if self.__color__ == "WHITE":
-            return "♜"
+            return "♝"
         else:
-            return "♖"
+            return "♗"