+++ game/queen.py
@@ -1,0 +1,2 @@
+from game.piece import Piece
+