+++ changelog.md
@@ -5,8 +5,45 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+#--------------Segundo Sprint----------------
+
+## [0.21] - 29-8-24
+
+
+## [0.20] - 27-8-24
+Added the real pieces to the board
+
+
+#--------------Primer Sprint----------------
+## [0.19] - 25-8-24
+Added methods permited_move to the pieces specifically
+
+## [0.18] - 24-8-24
+Added the permited moves for pawns including the moves when the pawn can eat another piece
+
+## [0.17] - 22-8-24
+Added exceptions to move_piece method in board and added the error of moving a piece where there is already a piece of the same color
+
+## [0.165] - 20-8-24
+FIxed complexity metods in board and added tests
+
+## [0.16] - 19-8-24
+Added function move_correct_color to chess, it prohibits the move of a piece that is not your color
+
+## [0.15] - 18-8-24
+Added permited moves to queen, king and bishop
+
+## [0.14] - 17-8-24
+Added function permited_move to board for Knight and Rook
+
+## [0.13] - 16-8-24
+Added classes pieces knight, bishop, queen and king. Added function show_board and the possibility to end game in any moment
+
+## [0.12] - 15-8-24
+Added function move to chess, move_piece method to board, and fixed get_piece method
+
 ## [0.11] - 14-8-24
-....
+Added get_piece method to board, and the pawns on the board
 
 ## [0.10] - 13-8-24
-....
+Added classes board, chess, and rook